requirements.txt
python-dotenv==1.0.0
streamlit==1.29.0
openai==1.4.0
beautifulsoup4==4.12.2
